+++ rslib/src/image_occlusion/image_occlusion_styling.css
@@ -19,7 +19,7 @@
     position: relative;
 }
 
-#image {
+#img {
     position: absolute;
     top: 0;
     left: 50%;
